svg/pservers/inheritance.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="8000"><title>Inheritance of SVG paint server properties</title><metadata><h:link href="https://svgwg.org/svg2-draft/pservers.html" rel="help"/><h:meta content="properties inherit according to the spec." name="assert"/><h:meta content="properties have expected initial values." name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/inheritance-testcommon.js"/><script><![CDATA[
assert_not_inherited('stop-color', 'rgb(0, 0, 0)', 'rgb(42, 53, 64)');
assert_not_inherited('stop-opacity', '1', '0.5');
  ]]></script></svg>
